Update supervisor to use getHostProvider
